service-cours/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cassandra-driver==3.28.0
httpx==0.26.0
uuid
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
aiohttp